Consolidate duplicate tool-evaluator branch tests into shared mixins
Enrich LogSafeSummary/AgentResponseReformat/ConversationHistoryReformat mixins and add a ToolResponseEvalUnitTests mixin (composed by tool_input_accuracy and tool_selection), then remove the per-file *InternalBranches tests they now cover and prune the newly-unused imports. Per-evaluator source coverage is unchanged (no regression).
